refactor: avatarの引数周りをリファクタ
