+++ frontend/src/components/environments/MainLayout/UserNavigation.tsx
@@ -2,7 +2,7 @@ import { FC, memo } from 'react';
 import { Button, Flex, Heading } from '@chakra-ui/react';
 import { useLogout, useProfile } from 'hooks/api';
 import { Link, useNavigate } from 'react-router-dom';
-import { UserAvatarContainer } from 'components/molecules/avatar/UserAvatarContainer';
+import { UserAvatarContainer } from 'components/organisms/avatar/UserAvatarContainer';
 
 export const UserNavigation: FC = memo(() => {
   const { logout } = useLogout();
@@ -17,7 +17,7 @@ export const UserNavigation: FC = memo(() => {
 
   return (
     <Flex p="5%" mt={4} align="center">
-      <UserAvatarContainer size="sm" src={user.avatarImageUrl} id={user.id} />
+      <UserAvatarContainer id={user.id} size="sm" src={user.avatarImageUrl} />
       <Flex flexDir="column" ml={4}>
         <Link to="/app/profile">
           <Heading as="h3" size="sm">
