+++ frontend/src/components/organisms/avatar/UserAvatar.tsx
@@ -2,19 +2,12 @@ import { FC, memo } from 'react';
 import { AvatarProps } from '@chakra-ui/react';
 import { LinkedAvatar } from 'components/atoms/avatar/LinkedAvatar';
 
-// type Props = {
-//   id: string;
-//   src: string;
-// };
 type Props = AvatarProps & {
-  id: string;
-  src: string;
+  link: string;
 };
 
 export const UserAvatar: FC<Props> = memo((props) => {
-  const { id, src, ...avatarProps } = props;
+  const { link, ...avatarProps } = props;
 
-  return (
-    <LinkedAvatar src={src} linkUrl={`/app/users/${id}`} {...avatarProps} />
-  );
+  return <LinkedAvatar link={link} {...avatarProps} />;
 });
