+++ frontend/src/components/organisms/avatar/UserAvatarContainer.tsx
@@ -6,22 +6,22 @@ import { UserAvatarWithBadge } from './UserAvatarWithBadge';
 
 type Props = AvatarProps & {
   id: string;
-  src: string;
 };
 
 export const UserAvatarContainer: FC<Props> = memo((props) => {
-  const { id, src, ...avatarProps } = props;
+  const { id, ...avatarProps } = props;
   const { users: friends } = useFriends();
   const { user } = useProfile();
   const isFriend = friends.find((friend) => friend.id === id) !== undefined;
   const isLoginUser = id === user.id;
+  const link = `/app/users/${id}`;
 
   return (
     <>
       {isFriend || isLoginUser ? (
-        <UserAvatarWithBadge id={id} src={src} {...avatarProps} />
+        <UserAvatarWithBadge id={id} link={link} {...avatarProps} />
       ) : (
-        <UserAvatar id={id} src={src} {...avatarProps} />
+        <UserAvatar link={link} {...avatarProps} />
       )}
     </>
   );
