+++ frontend/src/components/organisms/avatar/UserAvatarWithBadge.tsx
@@ -5,16 +5,16 @@ import { LinkedAvatar } from 'components/atoms/avatar/LinkedAvatar';
 
 type Props = AvatarProps & {
   id: string;
-  src: string;
+  link: string;
 };
 
 export const UserAvatarWithBadge: FC<Props> = memo((props) => {
-  const { id, src, ...avatarProps } = props;
+  const { id, link, ...avatarProps } = props;
   const { isOnline } = useFriendStatus(id);
   const badgeColor = isOnline ? 'green.500' : 'gray';
 
   return (
-    <LinkedAvatar src={src} linkUrl={`/app/users/${id}`} {...avatarProps}>
+    <LinkedAvatar link={link} {...avatarProps}>
       <AvatarBadge boxSize="1.1em" bg={badgeColor} />
     </LinkedAvatar>
   );
