+++ frontend/src/features/game/routes/GameTop.tsx
@@ -2,7 +2,7 @@ import { memo, FC, useContext } from 'react';
 import { Box, Button, Center, Text } from '@chakra-ui/react';
 import { Link } from 'react-router-dom';
 import { ContentLayout } from 'components/ecosystems/ContentLayout';
-import { UserAvatarContainer } from 'components/molecules/avatar/UserAvatarContainer';
+import { UserAvatarContainer } from 'components/organisms/avatar/UserAvatarContainer';
 import { OnlineUsersContext } from 'providers/OnlineUsersProvider';
 
 export const GameTop: FC = memo(() => {
