+++ frontend/src/features/profile/components/UserInfoCard.tsx
@@ -1,7 +1,7 @@
 import { memo, FC } from 'react';
 import { Flex, Text } from '@chakra-ui/react';
 import { User } from '@prisma/client';
-import { UserAvatarContainer } from 'components/molecules/avatar/UserAvatarContainer';
+import { UserAvatarContainer } from 'components/organisms/avatar/UserAvatarContainer';
 import { BlockButton } from './BlockButton';
 import { ProfileSetting } from './ProfileSetting';
 
@@ -26,10 +26,10 @@ export const UserInfoCard: FC<UserInfoCardProps> = memo(
         align="center"
       >
         <UserAvatarContainer
-          size="2xl"
-          name={user.nickname}
-          src={user.avatarImageUrl}
           id={user.id}
+          name={user.nickname}
+          size="2xl"
+          src={user.avatarImageUrl}
         />
         <Text fontSize="md" fontWeight="bold" pt="2">
           {user.nickname}
