+++ frontend/src/features/profile/components/matchhistory/AvatarWithName.tsx
@@ -1,6 +1,6 @@
 import { memo, FC } from 'react';
 import { Box, Text, VStack } from '@chakra-ui/react';
-import { UserAvatarContainer } from 'components/molecules/avatar/UserAvatarContainer';
+import { UserAvatarContainer } from 'components/organisms/avatar/UserAvatarContainer';
 
 type Props = {
   name: string;
@@ -14,7 +14,7 @@ export const AvatarWithName: FC<Props> = memo((props) => {
   return (
     <Box h="90" p={2}>
       <VStack>
-        <UserAvatarContainer size="md" id={id} src={avatarImageUrl} />
+        <UserAvatarContainer id={id} size="md" src={avatarImageUrl} />
 
         {/* TODO:ニックネームが長くなるときどう表示するか考える。暫定で省略している */}
         <Text maxWidth={20} noOfLines={1}>
